Better way of checking which side
