+++ src/main/java/frc/robot/util/AllianceFlipUtil.java
@@ -9,7 +9,6 @@ package frc.robot.util;
 
 import edu.wpi.first.math.geometry.*;
 import edu.wpi.first.wpilibj.DriverStation;
-import frc.robot.FieldConstants;
 
 public class AllianceFlipUtil {
   public static double applyX(double x) {
